etc:파일 삭제 및 함수 위치 이동
<description>
파일(ft_flush_pipe.c)삭제 및 함수 위치 (ft_flush_pipe.c -> ft_handle_pipe_two.c)이동
<issue-number>
